backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.12.1
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
validators==0.22.0
playwright==1.40.0
asyncio-throttle==1.0.2
python-multipart==0.0.6
websockets==12.0
redis==5.0.1
aiofiles==23.2.1